Embedded.Penetrations [release v.2.11.1] - отключена проверка на пересечения проходок в групповом режиме, на время анализа правильности работы алгоритма, его возможной оптимизации и для ускорения текущей операционной деятельности;

+++ Shared.Penetrations/PenetrLeaderPrimitiveCmd.cs
@@ -76,7 +76,7 @@ public class PenetrLeaderPrimitiveCmd : BCOM.IPrimitiveCommandEvents
                 1/ElementHelper.getActiveAnnotationScale());
 
             // ! перед добавлением в модель
-            mdlCell_setIsAnnotation(leader.MdlElementDescrP(), 1); 
+            mdlCell_setIsAnnotation((int)leader.MdlElementDescrP(), 1); 
             //leader.AsCellElement().ScaleUniform();
